// nightrider-mvp/nightrider-mvp/src/sre-agent/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { Ai, Annotation, App, Logger, MRNObject, ServiceStub, SqlDatabase, Tracer,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AI: Ai;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  INCIDENTS_DB: SqlDatabase;
  logger: Logger;
  TOOLS_API: ServiceStub<import('../tools-api/index.js').default>;
  tracer: Tracer;
}